requirements.txt
dnspython==2.7.0
exceptiongroup==1.2.2
iniconfig==2.0.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tomli==2.2.1
tzdata==2024.2
matplotlib==3.10.0
